small changes in structure and styling, added empty state message, improve blockList in settings

+++ Backend/controllers/post.js
@@ -121,11 +121,16 @@ export const deletePost = async (req,res)=>{
     try {
         const postId = req.params.id ;
 
+        if(!postId){
+            return res.status(400).json({err:"Post id is not provided."})
+        }
+
+        await SavedPosts.deleteMany({post:postId}) ;
         await Post.deleteOne({_id:postId});
-        await SavedPosts.findOneAndDelete({post:postId}) ;
+        
         return res.status(200).json({message:'Post deleted successfully.'});
     } catch (error) {
-        return res.status(500).json({message:"Internal server error."}) ;
+        return res.status(500).json({err:"Internal server error."}) ;
     }
 }
